FIX: update triangulations and reorderings
