+++ mantra/manifold_types.py
@@ -1,6 +1,5 @@
 from enum import Enum
 
-
 class Manifold2Type(Enum):
     """
     This represents the underlying 2-manifold type,
@@ -18,6 +17,7 @@ class Manifold2Type(Enum):
     P4_T_2 = "#^4 T^2"
     P5_T_2 = "#^5 T^2"
     P6_T_2 = "#^6 T^2"
+    P7_T_2 = "#^7 T^2"
     P8_T_2 = "#^8 T^2"
 
     # Non-orientable surfaces: connected sums of projective planes,
@@ -30,8 +30,12 @@ class Manifold2Type(Enum):
     P6_RP_2 = "#^6 RP^2"
     P7_RP_2 = "#^7 RP^2"
     P8_RP_2 = "#^8 RP^2"
+    P9_RP_2 = "#^9 RP^2"
     P10_RP_2 = "#^10 RP^2"
+    P11_RP_2 = "#^11 RP^2"
     P12_RP_2 = "#^12 RP^2"
+    P13_RP_2 = "#^13 RP^2"
+    P14_RP_2 = "#^14 RP^2"
     P15_RP_2 = "#^15 RP^2"
     P16_RP_2 = "#^16 RP^2"
     P17_RP_2 = "#^17 RP^2"
