+++ scripts/calculate_statistics.py
@@ -1,9 +1,37 @@
 import json
 import sys
 
+import matplotlib.pyplot as plt
 import pandas as pd
 import seaborn as sns
-import matplotlib.pyplot as plt
+import numpy as np
+from collections import defaultdict
+
+#TODO: Looka at what this does
+def print_statistics(dataset):
+    """Print per-class statistics of a dataset.
+
+    Parameters
+    ----------
+    dataset : list of dict
+        Dataset entries.
+    """
+    class_entries = defaultdict(list)
+    for entry in dataset:
+        class_entries[entry["name"]].append(entry)
+
+    print(
+        f"{'Class':<30} {'Count':>8} {'Min V':>8} {'Mean V':>8} {'Max V':>8}"
+    )
+    print("-" * 56)
+    for name in sorted(class_entries.keys()):
+        entries = class_entries[name]
+        nverts = [e["n_vertices"] for e in entries]
+        print(
+            f"{name:<30} {len(entries):>8} "
+            f"{min(nverts):>8} {round(np.mean(nverts),2):>8} {max(nverts):>8}"
+        )
+    print(f"\nTotal: {len(dataset)}")
 
 if __name__ == "__main__":
     with open(sys.argv[1]) as f:
@@ -22,6 +50,7 @@ if __name__ == "__main__":
 
     # So sue me, right?
     interesting_manifolds = ["RP^2", "T^2", "Klein bottle", "S^2"]
+    print(f"All manifold types: {df['name'].unique()}")
 
     df = df.query("name in @interesting_manifolds")
     df = df.groupby(["name", "n_vertices"]).size().to_frame("count")
